Push fixture requests down into the blackbox "Config" object
While trying to implement multi-backend integration testing, I found
it difficult to have a consistent interface in .patch, __init__, et al
to take care of fixture objects that are different for each backend.

Given that "request" was already pushed down into the "Config" object,
leverage it for all necessary fixtures and metadata rather than
splitting that responsibility in the "config" fixture.